add table read benchmark
